reuse code for unclaimed blocks rebuilding, refactoring
