+++ src/mzdiff.cpp
@@ -171,7 +171,7 @@ int main(int argc, char *argv[]) {
     try {
         Executable exeBase = loadExe(baseSpec, loadSeg, opt);
         Executable exeCompare = loadExe(compareSpec, loadSeg, opt);
-        RoutineMap map;
+        CodeMap map;
         if (!pathMap.empty()) {
             map = {pathMap, loadSeg};
         }
